% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_langs.R
\name{get_child_langs}
\alias{get_child_langs}
\title{Get the languages that are children of a parent vector}
\usage{
get_child_langs(dataset, query_vector)
}
\arguments{
\item{dataset}{character vector of the data set name (e.g., "CA16")}

\item{query_vector}{character vector of parent vector from which to retrieve child languages (e.g., "v_CA16_815")}
}
\value{
a tidy data frame with two columns: 1) parent vector label, and 2) child languages
}
\description{
Get the languages that are children of a parent vector
}
\examples{
\dontrun{
library(canlang)
options(cancensus.api_key = "your_api_key")
get_child_langs("CA16", "v_CA16_566")
}
}
